requirements.txt
flask
flask-socketio
requests
openai
faiss-cpu
langchain

